backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
azure-cognitiveservices-speech==1.40.0
openai==1.50.0
supabase==2.9.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
stripe==10.12.0
python-multipart==0.0.12
httpx==0.27.0
